# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IMDB REST API",
        "version": "1.0",
        "description": "Estoy muy feliz, es mi primera Rest Api!"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Development"
        }
    ],
    "paths": {
...
        "/users/:id": {
            "delete": {
                "security": [
                    {
                        "jwt_token": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "description": "Sirve para eliminar un usuario con su id especificado.",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Registro eliminado"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "jwt_token": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del usuario que deseamos actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "description": "Actualiza un registro con su id especificado.",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Users"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro actualizado"
                    }
                }
            },
            "get": {
                "security": [
                    {
                        "jwt_token": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del usuario a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "description": "Busca un usario según su id.",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Registro encontrado"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "jwt_token": []
                    }
                ],
                "description": "Sirve para obtener todos los usuarios de la BD",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina a obtener",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limite de datos por pagina",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registros encontrados"
                    }
                }
            },
            "post": {
                "description": "Sirve para crear un nuevo usuario.",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Users"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registro creado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Users": {
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "reset_token": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "jwt_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}